Refactoring of data workflow.

+++ Unite.Genome.Annotations/Clients/Ensembl/Resources/Vep/AnnotatedVariantResource.cs
@@ -14,5 +14,5 @@ internal record AnnotatedVariantResource
     public AffectedTranscriptResource[] AffectedTranscripts { get; set; }
 
     [JsonIgnore]
-    public long VariantId => long.Parse(Id.Split(".").First());
+    public int VariantId => int.Parse(Id.Split(".").First());
 }
